update cheat sheet

+++ dev/builder.py
@@ -4,7 +4,7 @@ folder_path = "../cheatsheet"
 
 for foldername, _, filenames in os.walk(folder_path):
     for filename in filenames:
-        if filename.endswith(".md"):
+        if filename.endswith(".md") and os.path.join(foldername, filename) == "README.md":
             os.remove(os.path.join(foldername, filename))
             # print(f"Файл удален: {os.path.join(foldername, filename)}")
